+++ src/main/java/com/bank/jbudget/repositories/CsvTransactionTagRepository.java
@@ -3,5 +3,5 @@ package com.bank.jbudget.repositories;
 import com.bank.jbudget.domain.CsvTransactionTag;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CsvTransactionTagRepository extends CrudRepository<CsvTransactionTag,Integer>{
+public interface CsvTransactionTagRepository extends CrudRepository<CsvTransactionTag, Long>{
 }
